+++ packages/keyring-api/src/v2/api/index.ts
@@ -6,3 +6,4 @@ export * from './keyring-snap-rpc';
 export * from './create-account';
 export * from './export-account';
 export * from './private-key';
+export * from './errors';