+++ packages/keyring-snap-sdk/src/v2/index.ts
@@ -1,1 +1,2 @@
 export * from './rpc-handler';
+export * from './errors';