+++ packages/keyring-eth-trezor/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Implement `deleteAccounts` method for bulk account deletion with best-effort semantics ([#637](https://github.com/MetaMask/accounts/pull/637))
+  - All accounts are attempted even if some fail, and a `DeleteAccountsError` is thrown if any deletion fails.
+
 ### Changed
 
 - Bump `@metamask/keyring-sdk` from `^3.0.0` to `^3.1.0` ([#614](https://github.com/MetaMask/accounts/pull/614))